+++ cmp/dataview/DataView.js
@@ -10,8 +10,7 @@ import {uses, hoistCmp} from '@xh/hoist/core';
 import {grid} from '@xh/hoist/cmp/grid';
 import {splitLayoutProps} from '@xh/hoist/utils/react';
 import {DataViewModel} from './DataViewModel';
-import {throwIf} from '@xh/hoist/utils/js';
-import {isNumber} from 'lodash';
+import {apiRemoved} from '@xh/hoist/utils/js';
 
 /**
  * A DataView is a specialized version of the Grid component. It displays its data within a
@@ -23,8 +22,9 @@ export const [DataView, dataView] = hoistCmp.withFactory({
     className: 'xh-data-view',
 
     render({model, className, ...props}) {
-        const [layoutProps, {rowCls, itemHeight, onRowDoubleClicked}] = splitLayoutProps(props);
-        throwIf(!isNumber(itemHeight), 'Must specify a number for itemHeight in DataView.');
+        apiRemoved(props.itemHeight, 'itemHeight', 'Specify itemHeight on the DataViewModel instead.');
+
+        const [layoutProps, {rowCls, onRowDoubleClicked}] = splitLayoutProps(props);
 
         return grid({
             ...layoutProps,
@@ -33,7 +33,7 @@ export const [DataView, dataView] = hoistCmp.withFactory({
             agOptions: {
                 headerHeight: 0,
                 rowClass: rowCls,
-                getRowHeight: () => itemHeight
+                getRowHeight: () => model.itemHeight
             },
             onRowDoubleClicked
         });
@@ -42,9 +42,6 @@ export const [DataView, dataView] = hoistCmp.withFactory({
 DataView.propTypes = {
     /** Primary component model instance. */
     model: PT.oneOfType([PT.instanceOf(DataViewModel), PT.object]),
-
-    /** Row height for each item displayed in the view */
-    itemHeight: PT.number.isRequired,
 
     /** CSS class used for each row */
     rowCls: PT.string,